修复：sendLocationMessage 经纬度参数改为 double 类型

+++ lib/jmessage_flutter.dart
@@ -2135,8 +2135,8 @@ class JMConversationInfo {
   }
   // sendLocation
   Future<JMLocationMessage> sendLocationMessage({
-    @required int latitude,
-    @required int longitude,
+    @required double latitude,
+    @required double longitude,
     @required num scale,
     String address,
     JMMessageSendOptions sendOption,
